pyproject.toml
[tool.poetry]
name = "docker-tests"
version = "0.1.0"
description = ""
authors = ["Alexandre Lima <alexandrepompeu@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
streamlit = "^1.32.2"
matplotlib = "^3.8.3"
yfinance = "^0.2.37"
google-generativeai = "^0.4.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
